관리자의 브랜드별 매출 목록보기 최종 수정 완료

+++ src/main/java/com/example/finalproject/domain/admin/AdminResponse.java
@@ -9,9 +9,9 @@ public class AdminResponse {
     public static class SalesListDTO {
         private AdminDTO admin;
         private Long orderItemPrice;
-        private Long fee;
+        private Double fee;
 
-        public SalesListDTO(Admin admin, Long orderItemPrice, Long fee) {
+        public SalesListDTO(Admin admin, Long orderItemPrice, Double fee) {
             this.admin = new AdminDTO(admin);
             this.orderItemPrice = orderItemPrice;
             this.fee = fee;
