+++ src/main/java/com/example/finalproject/domain/order/Order.java
@@ -27,14 +27,14 @@ public class Order {
 
     private Integer totalOrderAmount; // 주문 전체 금액
 
-    private Integer fee; // 주문 전체 수수료
+    private Double fee; // 주문 전체 수수료
 
     @CreationTimestamp
     private Timestamp orderDate; // 주문일자
 
 
     @Builder
-    public Order(Integer id, User user, Delivery delivery, Integer totalOrderAmount, Integer fee, Timestamp orderDate) {
+    public Order(Integer id, User user, Delivery delivery, Integer totalOrderAmount, Double fee, Timestamp orderDate) {
         this.id = id;
         this.user = user;
         this.delivery = delivery;
